requirements.txt
ultralytics==8.3.123
torch==2.7.0
opencv-python==4.10.0.84
pyyaml==6.0.2
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
lap==0.5.12